Create README.md
# Tic Tac Toe Game

A classic Tic Tac Toe game built using HTML, CSS, and JavaScript. Play against a friend on the same device and aim for three in a row to win. Enjoy the timeless fun of Tic Tac Toe with an intuitive user interface.

**Features:**
- Local multiplayer mode.
- Win and draw detection.
- Easy-to-use interface.

**Technologies:**
HTML, CSS, JavaScript.

**Get Started:**
1. Open `index.html` in a web browser.
2. Click on empty cells to make your move.
3. Aim for a winning combination of three symbols in a row.

**Future Enhancements:**
- Two-player mode 

**License:**
This project is licensed under the MIT License.

---
**Note:** Insert a screenshot of the game interface here.